fix change in resource limits
