--- apps/api/src/app/rate-limiting/usecases/get-api-rate-limit-service-maximum-config/index.ts
@@ -1,1 +1,0 @@
-export * from './get-api-rate-limit-service-maximum-config.usecase';
