+++ apps/api/src/app/rate-limiting/usecases/index.ts
@@ -1,5 +1,4 @@
 import { GetApiRateLimitMaximum } from './get-api-rate-limit-maximum';
-import { GetApiRateLimitServiceMaximumConfig } from './get-api-rate-limit-service-maximum-config';
 import { EvaluateApiRateLimit } from './evaluate-api-rate-limit';
 import { GetApiRateLimitAlgorithmConfig } from './get-api-rate-limit-algorithm-config';
 import { GetApiRateLimitCostConfig } from './get-api-rate-limit-cost-config';
@@ -7,7 +6,6 @@ import { EvaluateTokenBucketRateLimit } from './evaluate-token-bucket-rate-limit
 
 export const USE_CASES = [
   //
-  GetApiRateLimitServiceMaximumConfig,
   GetApiRateLimitMaximum,
   GetApiRateLimitAlgorithmConfig,
   GetApiRateLimitCostConfig,
